third_party/WebKit/LayoutTests/editing/selection/mouse/drag-user-select-all-textarea.html

<!doctype HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
    textarea {
        font-family: Ahem;
        font-size: 10px;
    }
</style>
<textarea id="textarea" style="-webkit-user-select:all">foobarbaz</textarea>
<script>
test(function(){
    assert_not_equals(window.eventSender, undefined, 'this test requires window.eventSender');
    eventSender.dragMode = false;
    eventSender.mouseMoveTo(textarea.offsetLeft + 30, textarea.offsetTop + 5);
    eventSender.mouseDown();
    eventSender.leapForward(100);
    eventSender.mouseMoveTo(textarea.offsetLeft + 60, textarea.offsetTop + 5);
    assert_equals(getSelection().toString(), 'bar');
    },
'User can select partially in -webkit-user-select:all textarea');
</script>
